Add serial number randomness requirement.
Statement of intent:

As demonstrated in
https://events.ccc.de/congress/2008/Fahrplan/attachments/1251_md5-collisions-1.0.pdf,
hash collisions can allow an attacker to forge a signature on the certificate
of their choosing. The birthday paradox means that, in the absence of random
bits, the security level of a hash function is half what it should be. Adding
random bits to issued certificates mitigates collision attacks and means
that an attacker must be capable of a much harder preimage attack. For a
long time the BRs have encouraged adding random bits to the serial number
of a certificate, and it is now common practice. This ballot makes that best
practice required, which will make the Web PKI much more robust against all
future weaknesses in hash functions. Additionally, it replaces "entropy" with
"CSPRNG" to make the requirement clearer and easier to audit, and clarifies
that serial number must be positive.

In Section 1.6.1 of the Baseline Requirements,

ADD

CSPRNG: A random number generator intended for use in cryptographic system.

In Section 7.1 of the Baseline Requirements,

REPLACE

"CAs SHOULD generate non-sequential Certificate serial numbers that exhibit at
least 20 bits of entropy."

WITH

"Effective September 30, 2016, CAs SHALL generate Certificate serial numbers
greater than zero (0) containing at least 64 bits of output from a CSPRNG."