Updates the effective date & EV confusion
This updates the Effective Date from March 2020 to April 2020.
Some CAs have raised concerns that their customers may be
concerned this requires potential technical changes to their
systems and that those may occur during holiday freezes.

This also addresses some loopholes that were identified, namely:
1) As worded, this would appear to allow the EVGs to supercede
   the BRs, which is not the intent.
2) As worded, this would allow a CA to never revalidate certain
   pieces of information, including the domain name, so long as
   the customer stayed with their current CA, rather than the
   intent, which was permitting the reuse of information within
   the specified period (398 days).

This does not address all loopholes, particularly those with
respect to reissuance of certificates that were revoked due to
the CA failing to adequately validate the information originally,
but that should be seen as a separate issue with the reliability
of the EV Guidelines.