Update CP/CPS and Audit Report requirements
This updates the requirement on the CP/CPS to make an explicit annual
versioning scheme required, as required by Mozilla.

Related, this updates the Audit Report requirements to align with the
current Microsoft and CCADB requirements, by removing the requirement
for CAs to document explicitly which CAs are out of scope. It also
aligns the language for "publicly accessible" by unhyphenating it,
as well as adding Microsoft's requirement that the Audit Report be
provided as a text-searchable PDF, in line with other industry's
requirements on electronic document submission.