Require revocation reason for intermediates/roots
Align with Microsoft's latest root program policy, which requires the
CA indicate the reason for revocation if the revoked certificate is
an intermediate certificate. To avoid terminology confusion, this makes
the requirement apply for Root CA and Subordinate CA Certificates
alike.

This incorporates RFC 5280's SHOULD NOT for unspecified (0) as a MUST
NOT. It also incorporates the prohibition on certificateHold, by way
of Section 4.9.13.

This clarifies for OCSP responses that the reason MUST be indicated
via the RevokedInfo. RFC 6960 permits any CRL entry extension to also
appear as an extension in the singleResponses field (see RFC 6960,
Section 4.4.5), thus it's possible to omit from the RevokedInfo and
instead stash as an extension