Initial draft to require public disclosure
Based on discussions on the 2020-02-27 Validation Subcommittee call,
this attempts to put in a basic requirement that the CA document
publicly what sources they use as a Registration Agency or
Incorporating Agency.

It is meant to be lightweight, while providing suitable transparency,
by merely requiring public documentation be made available what
source(s) will be used and what the allowed value(s) are for those
sources. The CA is not required to place this information in their
CP/CPS, but is required to place a pointer to this information so that
it can be discovered. Further, the CA is required to provide a basic
changelog that MUST minimally include a version and date, to avoid any
ambiguities when referencing this data.