Tweak the Name Encoding Rules
Address the comments from SecureTrust by making it clearer that the
requirements apply to both IDNA2003 and/or IDNA2008, by
incorporating RFC 8399.

Address the comments from Mozilla by clarifying that as far as the
Baseline Requirements go, this is not intended to require
revocation of existing, non-TLS-capable CAs, but is intended to
ensure that all newly issued certificates comply with this
requirement. Any CA which has a non-compliant configuration may
need to transition to a new issuing intermediate, and thus an
effective date of 2020-09-30 is provided to allow for key
ceremonies to be performed.