Shuffle Delegated Responder Clarifications
In the process of incorporating Microsoft's requirements
on EKUs, the requirement was worded in a way that clarified
the existing RFC 6960 requirements (and those of Section
4.9.9 of the BRs).

This caused confusion for some, thinking it was a new
requirement being introduced as part of a browser
alignment, but it's actually an existing requirement that
goes back to v1 of the BRs.

To resolve this confusion, move this requirement to the
"Cleanups and Clarifications" ballot, to make sure that it
is more prominent for CAs.