Update section links
As of
https://github.com/cabforum/build-guidelines-action/releases/tag/v2.0.0-rc3
section links now include the section number (e.g. "1.1 Section One"
becomes "#11-section-one"). This makes it easier to review and maintain
links, as any re-organizations will cause the links to break, which the
continuous integration will warn on.

This change rewrites all of the internal links to use the new style, and
bumps the dependency to rc3 (which will hopefully be the last release).