Harmonize styles further
This is a series of small list formatting cleanups to ensure proper
display. However, it does renumber Section 8.5.5 of the EVGs to use
Arabic numerals for the requirements, aligning with the other
requirements in Section 8.5, instead of using capitalized letters for
the list. There are no internal references to those requirements, and
so this should be safe.