Remove stray footnote
During the SC26 cleanup work, there was a stray footnote that got
left in which rendered incorrectly, relating to Section 3.2.2.6 and
wildcards. This correctly removes it.

Although we do support footnotes in the PDF and Docx writers, and
we do use them, Section 3.2.2.6 was modified during SC26 to make the
normative requirement clearer and part of the text, rather than as
a footnote.