Update the SCWG to use a Docker image for CI
This moves from using a GitHub Action as a Docker container that is
built on demand to using a public Docker container image hosted on
GitHub Container Registry (GHCR).

Prior to this change, GitHub Workflows check out the Action
repository and build the Docker image on demand for the workflow.
The downside to this is the external dependencies that are loaded
on demand (such as fonts and TeX Live packages) are subject to the
whims of the network gods and solar flares, and thus can result in
failed CI runs or very long (1+ hour) runs.

With this change, this switches to use the Docker image that is
automatically published to GHCR whenever we commit to the
cabforum/build-guidelines-action repository, which has all of the
resources integrated. As this now only depends on GitHub's network,
it's more reliable. As a bonus, it also executes much faster.