Introduce Precertificate Signing CA Profile
A Precertificate Signing CA, ontologically, a type of Technically
Constrained Non-TLS Sub-CA, by virtue of the Extended Key Usage. To
avoid ambiguity, this introduces a profile specific for Precert Signing
CAs that make it clear that the Precert Signing EKU MUST be the only EKU
present, to avoid a situation similar to that seen with OCSP responders.

RFC 6962 is somewhat fast and loose with respect to whether or not
"CA:true" is required in the profile for these, but in practice,
implementations of logs, and existing CAs, do expect CA:true.

Although not meant to be a normative change from the existing practices
and consequences of existing requirements, it does make explicit that
such CAs MUST only sign Precertificates; although this is less critical
given the EKU constraint (to being a singular EKU), it represents a
defense in depth approach consistent with existing practice.