Suggested changes for CRLs and OCSP
The current text proposes moving 12 months to 13 months (398 days). This instead aligns with 366 days as the SHOULD (the number of days in a leap year), with a SHALL NOT of 367 days (i.e. to account for leap seconds)

This moves the definition of Day and Year from being specific to Validity Internal and instead unambiguously moves them into conventions, 1.6.4, which is also where times and timezones are declared.

This removes the existing OCSP transition text, because we're now after 2020-09-30. For the CRL text, this ensures that the conditional clauses appear on the same line as their condition, to avoid the risk of subsequent paragraphs being interpreted as scoped to the condition (i.e. that there are no rules for Subordinate CA Certificates prior to 2022-02-01).