Precertificates and Precertificate Signing CAs
This introduces the notion of Precertificates and Precertificate
Signing CAs as part of the Profiles, and captures the existing
requirements from RFC 6962. It defines a Precertificate as based
on a transformation of an existing Certificate conforming to one
of the profiles, as opposed to attempting to define variants for
every version or how to construct a Precertificate for a given
profile.

This attempts to similarly capture that, for purposes of compliance,
a Precertificate is treated as if there is an equivalent Certificate,
by reflecting that Precertificates need to match a Certificate based
on the transformations defined, and that the Certificates need to
match the profiles defined.