Add DNSSEC validation requirement in 3.2.2.8
In order to enable domain owners that want to ensure their use of DNSSEC is protecting their domain as assumed by the security properties of DNSSEC, CAs should be required to validate DNSSEC. While this would ideally occur for all DNS queries, the most impactful and critical record in the certificate issuance process for this type of validation is the CAA record. Thus by adding a requirement to validate DNSSEC chains in Section 3.2.2.8 (which is somewhat already implied within this section as being expected already), we can ensure the ecosystem as a whole benefits from the interaction between the security properties of DNSSEC and TLS certificate issuance.