Clarify Authorization Domain Name
Simplify the definition of ADN and move the algorithm in 3.2.2.4.

Define Base Domain Name as a function of any given FQDN.

Move the definition of Domain Contact into the only remaining method that uses it.

Simplify the language in each method around pruning labels and wildcard
issuance. Add corresponding language about following CNAMEs when choosing the
ADN, and around onion issuance.

Use "ADN" consistently in method definitions instead of "FQDN" (which is
insufficiently precise as to whether it's the applied-for FQDN or the ADN).

Clean up the language in the onion section around ADNs.

Replace retired methods with language saying they are retired.